add strawberry server
